Query database by item_id for single row return
The code used to use the item_id as an index into the list of all
catalog items when returning a single item.  This worked as long as the
item_id's were sequential and started at 1.  The new code does uses a
WHERE clause on the database query to return the specified item id.